fix(a11y): render <html lang={locale}> by moving html/body to [locale] layout (#1617)
Every page on every tenant served <html> with no `lang` attribute,
failing WCAG A and blocking screen-reader language detection. The
root `app/layout.tsx` rendered <html> but sat above the [locale]
segment, so there was no static way to know the user's language
there.

Canonical next-intl pattern: move <html>/<body> into the locale
layout (which has `params.locale` available) and make the root
layout a pass-through. Next.js 15 allows this as long as exactly
one layout in the tree emits <html>/<body>. The standalone
`global-error.tsx` already emits its own <html>/<body> so error
paths still work.

- `app/layout.tsx` → minimal `return children;` pass-through.
- `app/[locale]/layout.tsx` → now owns <html lang={locale}>,
  <body>, font variables, generateMetadata, and the existing
  provider chain (intl / posthog / query / settings) +
  InstallPrompt + ServiceWorkerRegister.

Regression guard: `e2e/html-lang.spec.ts` asserts `document
  documentElement.lang` equals "fr" on /fr/courses and "en" on
/en/courses.

Fixes #1617
